Remove no directory error

+++ scraper.py
@@ -1,6 +1,7 @@
 import twint
 import sys
 import shutil
+import os
 
 c = twint.Config()
 
@@ -8,6 +9,8 @@ c.Username = sys.argv[1]
 c.Limit = 20
 c.Store_json = True
 c.Output = "tweets"
-shutil.rmtree("tweets")
+
+if os.path.exists("tweets/tweets.json"):
+    shutil.rmtree("tweets")
 
 twint.run.Search(c)
